refactor: Add config folder with url, env check, some smaller restructuring
