+++ src/helpers/convertUnits.ts
@@ -3,5 +3,5 @@
  * @param kelvin temperature in kelvin
  */
 export function kelvinToCelsius(kelvin: number): string {
-  return `${(kelvin - 273.15).toFixed()} °C`;
+  return `${(kelvin - 273.15).toFixed()}°C`;
 }
